+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './StrapiImage';
+export * from './StrapiImageRenderer';
 export * from './StrapiImageConfigProvider';