+++ vite.config.ts
@@ -2,11 +2,13 @@ import { resolve } from 'path';
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import dts from 'vite-plugin-dts';
+import { libInjectCss } from 'vite-plugin-lib-inject-css';
 import { name } from './package.json';
 
 export default defineConfig({
   plugins: [
     react(),
+    libInjectCss(),
     dts({
       tsconfigPath: resolve(__dirname, 'tsconfig.app.json'),
       exclude: ['**/*.test.*', '**/stories/**/*'],